examples/python/with_proto/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.6
setuptools==65.6.3
